feat: add POST /v1/ingest endpoint for server-side ingestion
Eliminates the Redb exclusive lock conflict between `corvia serve` and
`corvia workspace ingest` by running the adapter pipeline server-side,
reusing the server's already-open stores. Closes #4.

- Extract shared ingest logic into `corvia-kernel::ingest` module
  (infer_content_role, infer_source_origin, wire_pipeline_relations,
  run_workspace_ingest) so both CLI and server can use it
- Add POST /v1/ingest (202 Accepted, background spawn_blocking task)
  with atomic check-and-set guard (409 if already running)
- Add GET /v1/ingest/status for polling (idle/running/completed/failed)
- Refactor CLI workspace.rs to delegate to kernel function (~660 lines removed)
- Handle RwLock poisoning, session_ingest_lock interaction, 10min timeout
