+++ crates/corvia-cli/src/main.rs
@@ -621,6 +621,10 @@ async fn cmd_serve() -> Result<()> {
     let config_path = CorviaConfig::config_path();
     let cluster_store = Arc::new(corvia_server::dashboard::clustering::ClusterStore::new());
     let (hook_sessions, _hook_rx) = corvia_server::dashboard::session_watcher::SessionWatcherState::new();
+    // Derive workspace root from config_path parent (not current_dir)
+    let workspace_root = config_path.parent()
+        .map(|p| p.to_path_buf())
+        .unwrap_or_else(|| std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from(".")));
     let state = Arc::new(corvia_server::rest::AppState {
         store, engine, coordinator, graph, temporal, data_dir,
         rag: Some(rag), ready: ready.clone(), default_scope_id,
@@ -630,6 +634,8 @@ async fn cmd_serve() -> Result<()> {
         gc_history: Arc::new(corvia_kernel::ops::GcHistory::new(50)),
         session_ingest_lock: tokio::sync::Mutex::new(()),
         hook_sessions: hook_sessions.clone(),
+        workspace_root,
+        ingest_status: Arc::new(std::sync::RwLock::new(corvia_kernel::ingest::IngestStatus::idle())),
     });
     // Background cluster recompute every 60s
     {
@@ -830,7 +836,7 @@ async fn cmd_ingest(path: Option<&str>, incremental: bool, files: &[String]) -> 
 
             // Wire graph edges from pipeline relations
             let edges = if !pipeline_relations.is_empty() {
-                wire_pipeline_relations(
+                corvia_kernel::ingest::wire_pipeline_relations(
                     &pipeline_relations, &chunks, &stored_ids, &*graph,
                 ).await
             } else {
@@ -931,7 +937,7 @@ async fn cmd_ingest(path: Option<&str>, incremental: bool, files: &[String]) -> 
 
         // Step 5: Wire relations from pipeline (now native to ChunkingStrategy)
         let relations_stored = if !pipeline_relations.is_empty() {
-            wire_pipeline_relations(
+            corvia_kernel::ingest::wire_pipeline_relations(
                 &pipeline_relations, &processed, &stored_ids, &*graph,
             ).await
         } else {
@@ -2304,96 +2310,6 @@ fn wire_doc_to_code_relations(
     }
 
     edges
-}
-
-/// Resolve pipeline relations and store them as graph edges.
-///
-/// Uses `(source_file, start_line)` to match relations to stored chunks.
-/// Best-effort: unresolvable cross-file references are silently skipped.
-pub(crate) async fn wire_pipeline_relations(
-    relations: &[corvia_kernel::chunking_strategy::ChunkRelation],
-    processed: &[corvia_kernel::chunking_strategy::ProcessedChunk],
-    stored_ids: &[uuid::Uuid],
-    graph: &dyn GraphStore,
-) -> usize {
-    let mut relations_stored = 0;
-    let mut source_miss = 0u64;
-    let mut target_miss = 0u64;
-    for rel in relations {
-        // Find the source chunk by (source_file, start_line) match
-        let from_idx = processed.iter().position(|pc| {
-            pc.metadata.source_file == rel.from_source_file && pc.start_line == rel.from_start_line
-        });
-        let from_uuid = match from_idx {
-            Some(idx) if idx < stored_ids.len() => stored_ids[idx],
-            _ => {
-                source_miss += 1;
-                tracing::debug!(
-                    source_file = %rel.from_source_file,
-                    start_line = rel.from_start_line,
-                    relation = %rel.relation,
-                    "wire_pipeline_relations: source chunk not found"
-                );
-                continue;
-            }
-        };
-
-        // Find the target chunk by file name (and optionally symbol name)
-        let to_uuid = processed.iter().zip(stored_ids.iter()).find_map(|(pc, id)| {
-            if pc.metadata.source_file == rel.to_file {
-                if let Some(ref name) = rel.to_name {
-                    if pc.content.contains(name) {
-                        return Some(*id);
-                    }
-                } else {
-                    return Some(*id);
-                }
-            }
-            None
-        });
-
-        if let Some(to_uuid) = to_uuid {
-            // Filter self-edges for "imports" and "calls" relations
-            if (rel.relation == "imports" || rel.relation == "calls") && to_uuid == from_uuid {
-                continue;
-            }
-            if let Err(e) = graph.relate(&from_uuid, &rel.relation, &to_uuid, None).await {
-                tracing::warn!("Failed to store relation: {e}");
-            } else {
-                relations_stored += 1;
-            }
-        } else {
-            target_miss += 1;
-            tracing::debug!(
-                target_file = %rel.to_file,
-                target_name = ?rel.to_name,
-                relation = %rel.relation,
-                "wire_pipeline_relations: target chunk not found"
-            );
-        }
-    }
-    if source_miss > 0 || target_miss > 0 {
-        let total = relations.len();
-        // Warn when less than half of relations resolved — signals adapter/version mismatch
-        if relations_stored * 2 < total {
-            tracing::warn!(
-                total,
-                stored = relations_stored,
-                source_miss,
-                target_miss,
-                "wire_pipeline_relations: most relations failed to resolve — check adapter version"
-            );
-        } else {
-            tracing::info!(
-                total,
-                stored = relations_stored,
-                source_miss,
-                target_miss,
-                "wire_pipeline_relations: relation wiring summary"
-            );
-        }
-    }
-    relations_stored
 }
 
 fn print_server_search_results(results: &[server_client::SearchResultDto], _show_namespace: bool) {
