+++ crates/corvia-cli/src/workspace.rs
@@ -1,9 +1,5 @@
 use anyhow::{Context, Result};
 use corvia_common::config::{CorviaConfig, InferenceProvider, RepoConfig, WorkspaceConfig};
-use corvia_common::constants::{CLAUDE_SESSIONS_ADAPTER, USER_HISTORY_SCOPE};
-use corvia_kernel::adapter_discovery;
-use corvia_kernel::chunking_pipeline::register_adapter_chunking;
-use corvia_kernel::process_adapter::ProcessAdapter;
 use std::path::{Path, PathBuf};
 
 /// Validate that a name is safe to use as a directory component.
@@ -327,95 +323,14 @@ pub fn remove_repo_from_config(config: &mut CorviaConfig, name: &str) -> Result<
 /// - Markdown in docs/ (fallback) → "design"
 /// - Code extensions (.rs, .py, .ts, etc.) → "code"
 /// - Config/other → None (unclassified)
-pub fn infer_content_role(source_file: &str) -> Option<String> {
-    let lower = source_file.to_lowercase();
-    let file_name = std::path::Path::new(source_file)
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or("");
-
-    // Special files: instruction role
-    if file_name == "AGENTS.md" || file_name == "CLAUDE.md" || file_name == "README.md" {
-        return Some("instruction".into());
-    }
-
-    // Memory directory
-    if lower.contains("/.memory/") || lower.starts_with(".memory/") || lower == ".memory" {
-        return Some("memory".into());
-    }
-
-    // Directory-based inference for markdown/docs
-    if lower.ends_with(".md") || lower.ends_with(".mdx") {
-        // Check specific subdirs first (more specific wins)
-        if lower.contains("/rfcs/") || lower.starts_with("rfcs/") {
-            return Some("design".into());
-        }
-        if lower.contains("/plans/") || lower.starts_with("plans/") {
-            return Some("plan".into());
-        }
-        if lower.contains("/decisions/") || lower.starts_with("decisions/") {
-            return Some("decision".into());
-        }
-        if lower.contains("/learnings/") || lower.starts_with("learnings/") {
-            return Some("learning".into());
-        }
-        // Fallback: any markdown in a docs/ tree
-        if lower.contains("/docs/") || lower.starts_with("docs/") {
-            return Some("design".into());
-        }
-        return None;
-    }
-
-    // Code file extensions
-    const CODE_EXTS: &[&str] = &[
-        "rs", "py", "js", "jsx", "ts", "tsx", "go", "java", "rb", "php",
-        "c", "cpp", "h", "hpp", "cs", "swift", "kt", "scala", "ex", "exs",
-        "sh", "bash", "zsh", "toml", "yaml", "yml", "json",
-    ];
-    let ext = std::path::Path::new(source_file)
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("");
-    if CODE_EXTS.contains(&ext) {
-        return Some("code".into());
-    }
-
-    None
-}
-
-/// Infer `source_origin` based on where the file came from.
-///
-/// - Repo ingestion (repo_name is Some): "repo:<name>"
-/// - .memory/ files: "memory"
-/// - Workspace docs (repo_name is None): "workspace"
-pub fn infer_source_origin(repo_name: Option<&str>, source_file: &str) -> Option<String> {
-    let lower = source_file.to_lowercase();
-    if lower.contains("/.memory/") || lower.starts_with(".memory/") {
-        return Some("memory".into());
-    }
-    if let Some(name) = repo_name {
-        return Some(format!("repo:{name}"));
-    }
-    Some("workspace".into())
-}
-
-/// Simple glob matching for blocked paths (supports trailing `*` only).
-fn blocked_path_match(pattern: &str, path: &str) -> bool {
-    if let Some(prefix) = pattern.strip_suffix('*') {
-        path.starts_with(prefix)
-    } else {
-        path == pattern
-    }
-}
+// Re-export shared utilities from kernel (used by cmd_ingest in main.rs)
+pub use corvia_kernel::ingest::infer_content_role;
+pub use corvia_kernel::ingest::infer_source_origin;
 
 /// Ingest all (or one) workspace repos with namespace isolation.
 ///
-/// For each repo in the workspace, discovers an adapter at runtime, spawns
-/// it as a process, streams source files via JSONL, runs them through the
-/// chunking pipeline, embeds, and stores entries with namespace isolation.
-///
-/// Also ingests workspace-level docs (docs/decisions/, docs/learnings/, etc.)
-/// when running a full workspace ingest (no repo filter).
+/// Creates its own store/graph/engine for standalone CLI use, then delegates
+/// to [`corvia_kernel::ingest::run_workspace_ingest`] for the actual pipeline.
 ///
 /// The `fresh` parameter is accepted but not yet used (future: delete existing
 /// entries before re-ingest).
@@ -426,571 +341,23 @@ pub async fn ingest_workspace(
 ) -> Result<()> {
     let config_path = root.join("corvia.toml");
     let config = CorviaConfig::load(&config_path)?;
-    let ws = config
-        .workspace
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("Not a workspace — no [workspace] section in corvia.toml"))?;
 
     let data_dir = root.join(&config.storage.data_dir);
     let (store, graph) = corvia_kernel::create_store_at_with_graph(&config, &data_dir).await?;
     let engine = corvia_kernel::create_engine(&config);
 
-    // Discover adapters once for the whole workspace
-    let extra_dirs = config
-        .adapters
-        .as_ref()
-        .map(|a| a.search_dirs.clone())
-        .unwrap_or_default();
-    let discovered = adapter_discovery::discover_adapters(&extra_dirs);
+    let progress = corvia_kernel::ingest::PrintProgress;
+    corvia_kernel::ingest::run_workspace_ingest(
+        root,
+        &config,
+        store,
+        graph,
+        engine,
+        repo_filter,
+        None, // no session lock in CLI mode
+        &progress,
+    ).await?;
 
-    if discovered.is_empty() {
-        anyhow::bail!(
-            "No adapters found. Install corvia-adapter-basic to PATH or \
-             ~/.config/corvia/adapters/. Adapter binaries must be named corvia-adapter-<name>."
-        );
-    }
-
-    let default_name = config.adapters.as_ref().and_then(|a| a.default.as_deref());
-
-    let repos_to_ingest: Vec<&RepoConfig> = if let Some(name) = repo_filter {
-        let repo = ws
-            .repos
-            .iter()
-            .find(|r| r.name == name)
-            .ok_or_else(|| anyhow::anyhow!("Repo '{}' not found in workspace config", name))?;
-        vec![repo]
-    } else {
-        ws.repos.iter().collect()
-    };
-
-    for repo_config in &repos_to_ingest {
-        let repo_path = resolve_repo_path(root, &ws.repos_dir, repo_config);
-        if !repo_path.exists() {
-            println!(
-                "  Skipping {} — not cloned (run corvia workspace init)",
-                repo_config.name
-            );
-            continue;
-        }
-
-        let repo_path_str = repo_path
-            .to_str()
-            .ok_or_else(|| anyhow::anyhow!("Invalid path for repo {}", repo_config.name))?;
-
-        // Resolve adapter for this repo
-        let explicit = config
-            .sources
-            .as_ref()
-            .and_then(|sources| {
-                sources.iter().find(|s| s.path == repo_path_str).and_then(|s| s.adapter.as_deref())
-            });
-        let adapter_info = adapter_discovery::resolve_adapter(repo_path_str, &discovered, explicit, default_name)
-            .ok_or_else(|| anyhow::anyhow!("No suitable adapter found for '{}'", repo_path_str))?;
-
-        println!(
-            "\nIngesting {} (namespace: {}, adapter: {})...",
-            repo_config.name, repo_config.namespace, adapter_info.metadata.name
-        );
-
-        // Spawn adapter process
-        let mut process = ProcessAdapter::new(
-            adapter_info.binary_path.clone(),
-            adapter_info.metadata.clone(),
-        );
-        process.spawn().map_err(|e| anyhow::anyhow!(e))?;
-
-        // Ingest source files via IPC
-        let source_files = process
-            .ingest(repo_path_str, &config.project.scope_id)
-            .map_err(|e| anyhow::anyhow!(e))?;
-
-        // Build chunking pipeline with adapter strategies
-        let mut pipeline = corvia_kernel::create_chunking_pipeline(&config);
-
-        // Register adapter's chunking extensions via IPC
-        if !adapter_info.metadata.chunking_extensions.is_empty() {
-            let adapter_arc = std::sync::Arc::new(std::sync::Mutex::new(
-                ProcessAdapter::new(adapter_info.binary_path.clone(), adapter_info.metadata.clone()),
-            ));
-            adapter_arc.lock().unwrap().spawn().map_err(|e| anyhow::anyhow!(e))?;
-            register_adapter_chunking(
-                pipeline.registry_mut(),
-                adapter_arc,
-                &adapter_info.metadata.chunking_extensions,
-            );
-        }
-
-        let (processed, pipeline_relations, report) = pipeline.process_batch(&source_files)?;
-        println!(
-            "  {} files → {} chunks ({} merged, {} split)",
-            report.files_processed, report.total_chunks,
-            report.chunks_merged, report.chunks_split
-        );
-
-        // Build lookup for adapter-provided SourceMetadata overrides
-        let src_meta_lookup: std::collections::HashMap<&str, &corvia_kernel::chunking_strategy::SourceMetadata> =
-            source_files.iter().map(|sf| (sf.metadata.file_path.as_str(), &sf.metadata)).collect();
-
-        // Convert ProcessedChunks to KnowledgeEntries, embed, and store
-        let entries: Vec<corvia_common::types::KnowledgeEntry> = processed
-            .iter()
-            .map(|pc| {
-                let src_meta = src_meta_lookup.get(pc.metadata.source_file.as_str());
-                let mut entry = corvia_common::types::KnowledgeEntry::new(
-                    pc.content.clone(),
-                    config.project.scope_id.clone(),
-                    pc.metadata.source_file.clone(),
-                );
-                entry.workstream = src_meta
-                    .and_then(|m| m.workstream.clone())
-                    .unwrap_or_else(|| repo_config.namespace.clone());
-                entry.metadata = corvia_common::types::EntryMetadata {
-                    source_file: Some(pc.metadata.source_file.clone()),
-                    language: pc.metadata.language.clone(),
-                    chunk_type: Some(pc.chunk_type.clone()),
-                    start_line: Some(pc.start_line),
-                    end_line: Some(pc.end_line),
-                    content_role: src_meta
-                        .and_then(|m| m.content_role.clone())
-                        .or_else(|| infer_content_role(&pc.metadata.source_file)),
-                    source_origin: src_meta
-                        .and_then(|m| m.source_origin.clone())
-                        .or_else(|| infer_source_origin(
-                            Some(&repo_config.name),
-                            &pc.metadata.source_file,
-                        )),
-                };
-                entry
-            })
-            .collect();
-
-        let total = entries.len();
-        let mut stored_ids: Vec<uuid::Uuid> = Vec::with_capacity(total);
-        let mut stored = 0;
-        for batch in entries.chunks(corvia_kernel::introspect::EMBED_BATCH_SIZE) {
-            let texts: Vec<String> = batch.iter().map(|e| e.content.clone()).collect();
-            let embeddings = engine.embed_batch(&texts).await?;
-
-            for (entry, embedding) in batch.iter().zip(embeddings) {
-                let mut entry = entry.clone();
-                entry.embedding = Some(embedding);
-                store.insert(&entry).await?;
-                stored_ids.push(entry.id);
-                stored += 1;
-            }
-            println!("  embedded and stored {}/{}", stored, total);
-        }
-
-        // Wire relations from pipeline
-        let relations_stored = if !pipeline_relations.is_empty() {
-            crate::wire_pipeline_relations(
-                &pipeline_relations, &processed, &stored_ids, &*graph,
-            ).await
-        } else {
-            0
-        };
-        if relations_stored > 0 {
-            println!("  {relations_stored} graph relations stored");
-        } else if processed.len() > 10 {
-            tracing::warn!(
-                chunks = processed.len(),
-                input_relations = pipeline_relations.len(),
-                "0 graph relations stored for {} chunks — check adapter version",
-                processed.len()
-            );
-        }
-
-        // Shutdown adapter
-        process.shutdown().map_err(|e| anyhow::anyhow!(e))?;
-
-        println!(
-            "  {} chunks stored for namespace '{}'",
-            stored, repo_config.namespace
-        );
-    }
-
-    // --- Ingest workspace docs (only on full ingest, not single-repo) ---
-    if repo_filter.is_none()
-        && let Some(docs_config) = ws.docs.as_ref()
-            && let Some(workspace_docs_dir) = &docs_config.workspace_docs {
-                let docs_path = root.join(workspace_docs_dir);
-                if docs_path.exists() && docs_path.is_dir() {
-                    let docs_path_str = docs_path
-                        .to_str()
-                        .ok_or_else(|| anyhow::anyhow!("Invalid workspace docs path"))?;
-
-                    let blocked = docs_config
-                        .rules
-                        .as_ref()
-                        .map(|r| r.blocked_paths.clone())
-                        .unwrap_or_default();
-
-                    let adapter_info = adapter_discovery::resolve_adapter(
-                        docs_path_str,
-                        &discovered,
-                        None,
-                        default_name,
-                    );
-
-                    if let Some(adapter_info) = adapter_info {
-                        println!(
-                            "\nIngesting workspace docs ({})...",
-                            workspace_docs_dir
-                        );
-
-                        let mut process = ProcessAdapter::new(
-                            adapter_info.binary_path.clone(),
-                            adapter_info.metadata.clone(),
-                        );
-                        process.spawn().map_err(|e| anyhow::anyhow!(e))?;
-
-                        let source_files = process
-                            .ingest(docs_path_str, &config.project.scope_id)
-                            .map_err(|e| anyhow::anyhow!(e))?;
-
-                        // Filter: only allowed subdirs, exclude blocked paths
-                        let allowed = &docs_config.allowed_workspace_subdirs;
-                        let source_files: Vec<_> = source_files
-                            .into_iter()
-                            .filter(|sf| {
-                                let path = &sf.metadata.file_path;
-                                // Must be in an allowed subdir (or no restriction)
-                                let in_allowed = allowed.is_empty()
-                                    || allowed.iter().any(|sub| {
-                                        path.starts_with(&format!("{sub}/"))
-                                            || path.starts_with(sub.as_str())
-                                    });
-                                // Must not match blocked paths
-                                let is_blocked = blocked
-                                    .iter()
-                                    .any(|bp| blocked_path_match(bp, path));
-                                in_allowed && !is_blocked
-                            })
-                            .collect();
-
-                        if source_files.is_empty() {
-                            println!("  No docs files to ingest.");
-                        } else {
-                            let pipeline =
-                                corvia_kernel::create_chunking_pipeline(&config);
-                            let (processed, pipeline_relations, report) =
-                                pipeline.process_batch(&source_files)?;
-
-                            println!(
-                                "  {} files → {} chunks ({} merged, {} split)",
-                                report.files_processed,
-                                report.total_chunks,
-                                report.chunks_merged,
-                                report.chunks_split
-                            );
-
-                            // Build lookup for adapter-provided SourceMetadata overrides
-                            let docs_meta_lookup: std::collections::HashMap<&str, &corvia_kernel::chunking_strategy::SourceMetadata> =
-                                source_files.iter().map(|sf| (sf.metadata.file_path.as_str(), &sf.metadata)).collect();
-
-                            let entries: Vec<corvia_common::types::KnowledgeEntry> =
-                                processed
-                                    .iter()
-                                    .map(|pc| {
-                                        let src_meta = docs_meta_lookup.get(pc.metadata.source_file.as_str());
-                                        let mut entry =
-                                            corvia_common::types::KnowledgeEntry::new(
-                                                pc.content.clone(),
-                                                config.project.scope_id.clone(),
-                                                pc.metadata.source_file.clone(),
-                                            );
-                                        entry.workstream = src_meta
-                                            .and_then(|m| m.workstream.clone())
-                                            .unwrap_or_else(|| "docs".to_string());
-                                        entry.metadata =
-                                            corvia_common::types::EntryMetadata {
-                                                source_file: Some(
-                                                    pc.metadata.source_file.clone(),
-                                                ),
-                                                language: pc.metadata.language.clone(),
-                                                chunk_type: Some(
-                                                    pc.chunk_type.clone(),
-                                                ),
-                                                start_line: Some(pc.start_line),
-                                                end_line: Some(pc.end_line),
-                                                content_role: src_meta
-                                                    .and_then(|m| m.content_role.clone())
-                                                    .or_else(|| infer_content_role(
-                                                        &pc.metadata.source_file,
-                                                    )),
-                                                source_origin: src_meta
-                                                    .and_then(|m| m.source_origin.clone())
-                                                    .or_else(|| infer_source_origin(
-                                                        None,
-                                                        &pc.metadata.source_file,
-                                                    )),
-                                            };
-                                        entry
-                                    })
-                                    .collect();
-
-                            let total = entries.len();
-                            let mut stored_ids: Vec<uuid::Uuid> =
-                                Vec::with_capacity(total);
-                            let mut stored = 0;
-                            for batch in
-                                entries.chunks(corvia_kernel::introspect::EMBED_BATCH_SIZE)
-                            {
-                                let texts: Vec<String> =
-                                    batch.iter().map(|e| e.content.clone()).collect();
-                                let embeddings = engine.embed_batch(&texts).await?;
-                                for (entry, embedding) in batch.iter().zip(embeddings) {
-                                    let mut entry = entry.clone();
-                                    entry.embedding = Some(embedding);
-                                    store.insert(&entry).await?;
-                                    stored_ids.push(entry.id);
-                                    stored += 1;
-                                }
-                                println!(
-                                    "  embedded and stored {}/{}",
-                                    stored, total
-                                );
-                            }
-
-                            if !pipeline_relations.is_empty() {
-                                let relations_stored = crate::wire_pipeline_relations(
-                                    &pipeline_relations,
-                                    &processed,
-                                    &stored_ids,
-                                    &*graph,
-                                )
-                                .await;
-                                if relations_stored > 0 {
-                                    println!(
-                                        "  {relations_stored} graph relations stored"
-                                    );
-                                }
-                            }
-                            // No zero-relations warning here — workspace docs
-                            // (markdown/text) are not expected to produce graph relations.
-
-                            println!(
-                                "  {} chunks stored for workspace docs",
-                                stored
-                            );
-                        }
-
-                        process.shutdown().map_err(|e| anyhow::anyhow!(e))?;
-                    } else {
-                        println!(
-                            "\n  Skipping workspace docs — no suitable adapter found \
-                             (install corvia-adapter-basic)"
-                        );
-                    }
-                }
-            }
-
-    // --- Ingest Claude Code session history (only on full ingest) ---
-    if repo_filter.is_none() {
-        // Check for a user-history scope in [[scope]] config
-        let session_scope = config
-            .scope
-            .as_ref()
-            .and_then(|scopes| scopes.iter().find(|s| s.id == USER_HISTORY_SCOPE));
-
-        if session_scope.is_some() {
-            let adapter_info = discovered.iter().find(|a| a.metadata.domain == CLAUDE_SESSIONS_ADAPTER);
-            if let Some(adapter_info) = adapter_info {
-                let home = std::env::var("HOME").unwrap_or_else(|_| "/root".into());
-                let sessions_dir = PathBuf::from(home)
-                    .join(".claude")
-                    .join("sessions");
-
-                if sessions_dir.is_dir() {
-                    println!("\nIngesting Claude Code session history...");
-
-                    let sessions_path = sessions_dir
-                        .to_str()
-                        .unwrap_or("~/.claude/sessions");
-
-                    let mut process = ProcessAdapter::new(
-                        adapter_info.binary_path.clone(),
-                        adapter_info.metadata.clone(),
-                    );
-                    process.spawn().map_err(|e| anyhow::anyhow!(e))?;
-
-                    let source_files = process
-                        .ingest(sessions_path, USER_HISTORY_SCOPE)
-                        .map_err(|e| anyhow::anyhow!(e))?;
-
-                    if source_files.is_empty() {
-                        println!("  No new sessions to ingest.");
-                    } else {
-                        // Build entries directly from adapter output — each SourceFile
-                        // is one session turn, already chunked by the adapter. Skipping
-                        // the ChunkingPipeline preserves the per-turn source_version
-                        // (e.g. "ses-abc:turn-1") which the pipeline would collapse to
-                        // just the file_path (session ID).
-                        let session_count = {
-                            let mut seen = std::collections::HashSet::new();
-                            source_files.iter().for_each(|sf| { seen.insert(&sf.metadata.file_path); });
-                            seen.len()
-                        };
-
-                        println!(
-                            "  {} sessions → {} turn entries",
-                            session_count,
-                            source_files.len()
-                        );
-
-                        let entries: Vec<corvia_common::types::KnowledgeEntry> =
-                            source_files
-                                .iter()
-                                .map(|sf| {
-                                    let mut entry =
-                                        corvia_common::types::KnowledgeEntry::new(
-                                            sf.content.clone(),
-                                            USER_HISTORY_SCOPE.to_string(),
-                                            sf.metadata.source_version.clone(),
-                                        );
-                                    entry.workstream = sf.metadata.workstream.clone()
-                                        .unwrap_or_default();
-                                    entry.metadata =
-                                        corvia_common::types::EntryMetadata {
-                                            source_file: Some(
-                                                sf.metadata.file_path.clone(),
-                                            ),
-                                            language: sf.metadata.language.clone(),
-                                            chunk_type: Some("session-turn".into()),
-                                            start_line: None,
-                                            end_line: None,
-                                            content_role: sf.metadata.content_role.clone(),
-                                            source_origin: sf.metadata.source_origin.clone(),
-                                        };
-                                    entry
-                                })
-                                .collect();
-
-                        // Build a parallel vec of parent_session_id for edge creation.
-                        // Only turn-1 entries carry this (set by the adapter).
-                        let parent_ids: Vec<Option<String>> = source_files
-                            .iter()
-                            .map(|sf| sf.metadata.parent_session_id.clone())
-                            .collect();
-
-                        let total = entries.len();
-                        let mut stored = 0;
-                        let mut stored_ids: Vec<uuid::Uuid> = Vec::with_capacity(total);
-                        for batch in
-                            entries.chunks(corvia_kernel::introspect::EMBED_BATCH_SIZE)
-                        {
-                            let texts: Vec<String> =
-                                batch.iter().map(|e| e.content.clone()).collect();
-                            let embeddings = engine.embed_batch(&texts).await?;
-                            for (entry, embedding) in batch.iter().zip(embeddings) {
-                                let mut entry = entry.clone();
-                                entry.embedding = Some(embedding);
-                                store.insert(&entry).await?;
-                                stored_ids.push(entry.id);
-                                stored += 1;
-                            }
-                            println!("  embedded and stored {}/{}", stored, total);
-                        }
-
-                        // Create spawned_by graph edges for subagent sessions.
-                        // For each entry with a parent_session_id, find the parent
-                        // session's first-turn entry (by file_path == parent session ID)
-                        // and create: subagent_turn1 --spawned_by--> parent_turn1.
-                        // Create spawned_by graph edges for subagent sessions.
-                        // Invariant: adapter emits exactly one entry per turn, so
-                        // source_files[i] maps to stored_ids[i] and parent_ids[i].
-                        debug_assert_eq!(
-                            source_files.len(),
-                            stored_ids.len(),
-                            "1:1 source-to-entry invariant violated in session ingest"
-                        );
-                        {
-                            // Build lookup: session_id → first stored entry UUID.
-                            // Relies on adapter emitting turns in order (turn 1 first),
-                            // so or_insert picks the turn-1 entry for each session.
-                            let mut session_first_entry: std::collections::HashMap<String, uuid::Uuid> =
-                                std::collections::HashMap::new();
-                            for (i, sf) in source_files.iter().enumerate() {
-                                session_first_entry
-                                    .entry(sf.metadata.file_path.clone())
-                                    .or_insert(stored_ids[i]);
-                            }
-
-                            let mut edges_created = 0u32;
-                            for (i, parent_id) in parent_ids.iter().enumerate() {
-                                if let Some(parent_sid) = parent_id {
-                                    let child_entry_id = stored_ids[i];
-                                    if let Some(&parent_entry_id) = session_first_entry.get(parent_sid) {
-                                        if let Err(e) = graph
-                                            .relate(
-                                                &child_entry_id,
-                                                "spawned_by",
-                                                &parent_entry_id,
-                                                None,
-                                            )
-                                            .await
-                                        {
-                                            eprintln!(
-                                                "  warning: failed to create spawned_by edge: {e}"
-                                            );
-                                        } else {
-                                            edges_created += 1;
-                                        }
-                                    } else {
-                                        // Parent was ingested in a prior run. Look up via
-                                        // source_version index (cross-batch resolution).
-                                        let parent_sv = format!("{parent_sid}:turn-1");
-                                        match store
-                                            .get_by_source_version(USER_HISTORY_SCOPE, &parent_sv)
-                                            .await
-                                        {
-                                            Ok(Some(parent_entry)) => {
-                                                if let Err(e) = graph
-                                                    .relate(
-                                                        &child_entry_id,
-                                                        "spawned_by",
-                                                        &parent_entry.id,
-                                                        None,
-                                                    )
-                                                    .await
-                                                {
-                                                    eprintln!(
-                                                        "  warning: cross-batch spawned_by edge failed: {e}"
-                                                    );
-                                                } else {
-                                                    edges_created += 1;
-                                                }
-                                            }
-                                            Ok(None) => {
-                                                eprintln!(
-                                                    "  note: parent session {parent_sid} not found — \
-                                                     spawned_by edge skipped (not yet ingested or GC'd)"
-                                                );
-                                            }
-                                            Err(e) => {
-                                                eprintln!(
-                                                    "  warning: cross-batch parent lookup failed: {e}"
-                                                );
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            if edges_created > 0 {
-                                println!("  {edges_created} spawned_by graph edges created");
-                            }
-                        }
-
-                        println!("  {} session turns stored", stored);
-                    }
-
-                    process.shutdown().map_err(|e| anyhow::anyhow!(e))?;
-                }
-            }
-        }
-    }
-
-    println!("\nWorkspace ingest complete.");
     Ok(())
 }
 
@@ -1150,6 +517,7 @@ pub async fn workspace_status(root: &Path) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use corvia_kernel::ingest::blocked_path_match;
 
     #[test]
     fn test_repo_name_from_url() {
